[ig4a] Implements group join notification handler plugin on Android.
Summary:
This implements the Android plugin to call 'create_group_thread' and verify if a given thread is instamadillo.

Two things are needed to enable the codegen:
- switching completion context to be the last param in the extension header
- redeclare the completion callback as a dependency to the plugin\

While the steps are confusing, having the codegen still beats writing the JNI conversion classes by hand.

Test Plan: Tested on Android, verified the 'create_group_thread' requests are sent as expected.

Differential Revision: D66787620

fbshipit-source-id: ca1610a84519dc1751031c94e9c1907a1da4249a