[antlir2][rpm] repodata action must run locally
Summary:
This doesn't work remotely because it downloads `cas_artifact`s that are using
a different `use_case` that is not available on RE.

Somehow, scuba claims that 1% of actions run on RE and succeed, but that should
be impossible. Either way, almost all of these are either served by the cache
or run locally.
https://fburl.com/scuba/buck2_actions/a95j7qh4

Test Plan: testhard

Reviewed By: epilatow

Differential Revision: D68028682

fbshipit-source-id: ad297ec611f25aa654ff8bbda48ac26ed989f308