[metalos][vm] emit vms for all kernels in bot_generated/kernel/kernels/versions.bzl
Summary:
We used to generate vms only for kernels that were available through
one of the kernel selection.
This prevented a klp to be tested for a kernel that was otherwise published in
versions.bzl: T216130600

This change uses the kernels in versions.bzl as source of truth.
To prevent CI from building tons of vm images, CI is skipped for this folder.

Also prevent building/testing the images from //metalos/vm

Test Plan:
Using `05090069a50492a8ac825d74a36fa6fd01c70eff`, where we would get this
failure:
{P1739564975}

After the diff, the build succeeds.

Reviewed By: wujj123456

Differential Revision: D70113339

fbshipit-source-id: 2a0ad0f421da9a290776cc94d55498b926b46fc4