[antlir] Add systemd Wants= dependency for virtio-notify
Summary:
This unit has a pretty odd setup, there is some context in the diff that set it up that way (D55843433).

I'm adding the dependency back, as a Wants= so it's not a pure revert of the previous one.

The current setup keeps retrying the unit until network-online.target is reached, and sshd is running.
However if nothing is explicitly asking for network-online.target, it is pruned from the tree and is never marked as reached -- even though all its dependencies are satisfied.

Test Plan: local antlir VM

Reviewed By: vmagro

Differential Revision: D70569936

fbshipit-source-id: 62c4c801e79ed295e294e0122b133ba61d87e6a0