[antlir2][sendstream] delete parent/child size test
Summary:
This test was testing idealized parent/child relationships where we installed
the same large file from outside the image layer environment in both the parent
and the child and expected the kernel to keep track of it.

This test is flaky, and I think it's because sometimes we end up with two
copies of the outer file that are not the same (they may even have the same
contents, but are not literally the same extents).

Since we never actually have this behavior in prod-destined images, just delete
the test.

Test Plan: testhard

Reviewed By: naveedgol

Differential Revision: D71994625

fbshipit-source-id: 0daa70b7210b06d801750fe5f9c5c371afb3f0c5