[antlir2][feature] remove antlir2_configured_deps
Summary:
There was an earlier attempt to strip the antlir configuration from most
`deps`, but it ended up breaking a lot of things and was logically reverted,
but not completely deleted.
This diff completes that deletion by moving the two use cases back to just
`deps`, which is a no-op change.

Test Plan: waitforsandcastle

Reviewed By: justintrudell

Differential Revision: D74024425

fbshipit-source-id: 8cbd76547cc860d897b9d22c3c46d51cc79e2946