[antlir2][plugins] make init_tracing more robust
Summary:
Previously, `init_tracing` just set the current global tracing dispatcher as
the default in the plugin when the plugin was loaded, but that is prone to
ordering issues if the global dispatcher is registered/changed after the plugin
is loaded, since it would not propagate down to the plugin.

This diff makes this setup more robust by implementing a new
`PluginForwardingSubscriber` that is set as the default dispatch for each
plugin at load time, and simply forwards to whatever dispatcher is the default
in the main process space at calling time.

Test Plan:
```
❯ buck2 test fbcode//antlir/antlir2/features/test_only_features/trace/...
Buck UI: https://www.internalfb.com/buck2/21f66631-0d6e-416f-a2fa-80acaab27311
Tests finished: Pass 4. Fail 0. Fatal 0. Skip 0. Build failure 0
```
https://www.internalfb.com/intern/testinfra/testrun/2814750052665951

Reviewed By: naveedgol

Differential Revision: D75007839

fbshipit-source-id: bc8df3428dc6bc2d1310a7abd79b1f292eb4ce9c