[antlir2][systemd] default to "unconfigured"
Summary:
Instead of defaulting to `stable`, default to not setting a configuration at
all.

This is a cleaner way to let packages and implicit layers defer the systemd
configuration to the first `(parent_)layer` that requests something specific.

This diff also renames that option from `"inherit-parent"` to `"unconfigured"`
to accurately reflect how it does not actually configure systemd to match the
`parent_layer` but instead leaves it unspecified (so in practice, appearing as
if `"stable"` was set, even if the parent was `"cd"`)

Test Plan: waitforsandcastle

Reviewed By: sergeyfd

Differential Revision: D75973123

fbshipit-source-id: 27a6b3c7d9b01b38436d470c0336ca8d9ce21e77