tee: Use python3 for all fbs/third-party/python deps
Summary:
Prior to this, even though we built TEE projects linking against the antlir distro version of python, we still relied on the python 3.12 headers.

While this ordinarily _should_ work, I don't want to run the risk of mismatched symbols between fbsource/third-party and whatever centos has, so let's just keep this consistent and use the same python3 libraries and headers throughout.

This also updates the python3 rpm_library rule to point to the actual correct python3.12.so instead of python3.9.so (which is what python3.so links to by default, whoops).

Test Plan:
CI

Rollback Plan:

Reviewed By: zsol

Differential Revision: D77013760

fbshipit-source-id: 4a50b987b4b96e45030a09f566a0c9557a6be08f