[antlir] python: `$(location` for interpreter
Summary:
It's a bit unfortunate really that this code works out in such a way as to *feel* complicated even though it's actually kind of not

As a reminder, there's a concept of the "antlir python interpreter." The antlir python interpreter is a semantically distinct python interpreter from the fbcode one. In practice that's because it might have different python versions, and there's more information about why that comes up [here](https://www.internalfb.com/code/fbsource/[5cbba5a46afa]/fbcode/buck2/platform/python/interpreters.bzl?lines=77-96), but for the sake of argument lets make it simple: Let's pretend for a second that the antlir python interpreter distinguished itself by having an additional global `antlir_print` function. That's the most "semantically distinct" an interpreter can be.

The `python_interpreter` target being defined here is just a target that means "build me a python interpreter that supports `antlir_print`". Like always, you should build it for use with the target platform. A correct - and the simplest - way to implement this `python_interpreter` would be to just point it at a `cpp_binary` which is a fork of cpython. Well, building cpython from source is obviously not simple, but it would make setting up this target very simple.

Instead of that, we use prebuilt interpreters. In general when you use prebuilt artifacts anywhere, you have to deal with the problem that you need to get different prebuilt artifacts depending on the target platform; your prebuilt artifact built for Linux normally can't be the same as the one built for Mac. All that we're doing in the big select on this target is implementing that - this is "find me an interpreter that implements `antlir_print` for the target platform that I asked for. If we're building something that's targeted at `build-for-distro`, the interpreter is found somewhere on PATH. If we're building targeted at fbcode, we need to use the `image_command_alias` thing to get it.

Sean is right that the reason that need to be able to build an interpreter that supports `antlir_print` for an fbcode platform in the first place is because we need to compile pycs at build time using that interpreter, but compared to everything else I wrote above, that's kind of a detail.

Test Plan:
Same as previous diff

```
> sl rebase -d master
> sl graft D76638257
> buck cquery fbcode//mode/opt --target-universe fbcode//tee/snp/image/examples/sidecar:cvm_sidecar_image  fbcode//antlir/distro/toolchain/python:interpreter -a interpreter

{
  "fbcode//antlir/distro/toolchain/python:interpreter (cfg:opt-linux-x86_64-clang19-no-san-opt-by-default#9125eb972e86a5b1)": {
    "interpreter": "python3.12",
    "interpreter_for_manifest": "python3.12",
    "interpreter_home_for_manifest": "/usr"
  },
  "fbcode//antlir/distro/toolchain/python:interpreter (fbcode//buck2/platform/execution:linux-x86_64#502119a3550f1f14)": {
    "interpreter": "$(location fbcode//antlir/distro/toolchain/python:interpreter--centos9--python3.12 (fbcode//buck2/platform/execution:linux-x86_64#502119a3550f1f14))",
    "interpreter_for_manifest": "python3.12",
    "interpreter_home_for_manifest": "/usr"
  }
}
```

But anyway, I'm pretty confident in this.

The reason it didn't appear to be wrong earlier is because *usually*, the exec platform of an exec platform will just be itself, but there's nothing guaranteeing that.

Rollback Plan:

Reviewed By: get9

Differential Revision: D77285694

fbshipit-source-id: 116745eec562e5cadea9400d4def643d5c1dccb3