[cli.metadata] add criticality(level=NOT_CRITICAL)
Summary: This is a followup to D77585316 and D77551078. This adds `criticality = cli.criticality(level=cli.level.NOT_CRITICAL)` to the existing CLIs. This might not be entirely correct, but the owners should aim to update this themselves. D77551078 took care of the TIER_0 / Critical ones we care about.

Test Plan:
Sandcastle

`clideployer diff` P1857920124

Rollback Plan:

Reviewed By: iulianmac

Differential Revision: D77450636

fbshipit-source-id: 8b38d1f2c38bd941f6f148bc44c578f0c46c6a51