[antlir2][rpm] remove pkgid and from_repo from fact
Summary:
`pkgid` is not the traditional pkgid that we are used to dealing with (which is
a checksum of the entire `.rpm` file) so it is not very useful.

`from_repo` is also not as useful as expected and is sometimes blank/null for
reasons that I haven't been able to determine.

Test Plan:
waitforsandcastle

Rollback Plan:

Reviewed By: justintrudell

Differential Revision: D78029186

fbshipit-source-id: 8374b7a27b4095625dd64ba34974505b2b4db55d