antlir: Remove ambiguous target_compatible_with
Summary: This got added as part of the safer-compatible-with effort, but `antlir//antlir/os:centos9` and `antlir//antlir/distro:build-for-distro` are ambiguous with each other, causing buck to fail to resolve this target compatible with. Let's remove build-for-distro since an os constraint is always set if we're building this target.

Test Plan:
CI

Rollback Plan:

Reviewed By: lisroach

Differential Revision: D78562829

fbshipit-source-id: a377b96b25cc789fc805bd20743926ff2acde77b