[antlir2][rpm] improve scriptlet-failure behavior on new dnf versions
Summary:
The new version of `dnf4` in ELN suddenly treats scriptlet failures as hard
errors. This is a departure from previous `dnf4` versions, and while I
completely agree with it, we unfortunately live in a world where we need to
install some critical RPMs that have shitty postscripts that fail.

The new `dnf4` behavior does not give us any way to continue to allowlist said
RPMs, so this diff takes a pragmatic approach to just include a bit more
clarification on the behavior, and also removes the currently failing ELN test
that is impossible to fix.

Test Plan:
```name="Before simply deleting the ELN test"
❯ buck test fbcode//antlir/antlir2/features/rpm/tests:allowlisted-failing-postscripts-dnf5
...
2025-08-04T16:53:36.398634Z TRACE compile:rpms: rpm: dnf-driver: TxError("Could not run transaction. There were scriptlet errors in the following rpms. Newer versions of dnf can consider this a hard failure. {'antlir2-failing-postscripts'}")
```

Rollback Plan:

Reviewed By: sergeyfd

Differential Revision: D79568132

fbshipit-source-id: cedce4a875d95686de9b4417e6b8073ef9817918