tee: antlir: Add nasm support to antlir distro toolchain
Summary:
This is needed to unblock cathode onboarding.

Trying to build a rule with .asm files fails because we try to compile these with clang today:
https://www.internalfb.com/buck2/75ccd69c-ba3a-4824-bc55-8086b22526c7

Like the nvcc_wrapper, let's use the buck tool asm_wrapper. This is pretty tightly coupled with the cxx rules which pass specific flags to nasm that get stripped by this wrapper, so we need to use this.

Test Plan:
Build rule with asm deps in tee mode, it works:
https://www.internalfb.com/buck2/7a0afc09-de49-4079-8a3a-38d53676ac09

Rollback Plan:

Differential Revision: D79868680

fbshipit-source-id: f911a40be5c6c474e72935369bc6ad7ad4cf3fd9