antlir: Add clang compiler constraints to base platform
Summary:
This is necessary so that we get clang-specific flags (like `-Wno-unknown-pragma`) added by the macro layer.

Note that we need to silence the unused typedef warning because glog has a bit of code that fires this. We removed this in tp2's glog in D923305, but since glog comes from a distro rpm it's kinda hard to change the source.

Test Plan:
CI

Orchestrator builds:
https://www.internalfb.com/buck2/f72d518e-238d-4285-a4e9-2575e653b62d

Rollback Plan:

Reviewed By: Rittzz

Differential Revision: D80320849

fbshipit-source-id: a363b9e7cf5cfc5a27ac56227ab465555db8347a