[antlir2][sendstream] skip directories when checking for "hardlinkability"
Summary:
BTRFS snapshots are not recursive. The traces of leftover subvolumes persist in
child layers as directories with `ino == 2`. This can cause weird packaging
issues with sendstreams produced with these layers.

Luckily the fix is easy: just ignore directories when doing this check
This should always have been the case because directories can never be
hardlinked, but live and learn. We arguably should fail the image compilation
if non-empty subvolumes are discovered (since the contents will be immediately
lost in the child layer/phase), but I will leave that to a follow-up diff.

To date the only image that seems to have been affected is
`gedge.metalos.rootfs`.

Test Plan:
```name="Re-package a known-bad subvolume"
❯ buck run fbcode//mode/opt fbcode//antlir/antlir2/antlir2_packager:antlir2-packager -- --spec ~/tmp/gedge/spec.json --out ~/tmp/gedge/new.sendstream --layer ~/tmp/gedge/bad-sendstream --rootless
```

```name="It does not exhibit the old hardlink problem"
❯ btrfs receive --dump -f ~/tmp/gedge/new.sendstream | grep var/facebook/tupperware/agent/control
```
{P1912972632}

Whereas before it looked more like this, with hardlinks being created to
directories (which then fails to be received)
{P1912973266}

Rollback Plan:

Reviewed By: nikitakoshikov

Differential Revision: D80738788

fbshipit-source-id: fb2cbe41055ff37a67491adcbe44f4b480f48266