tee: Customize glog with fbcode-specific bits
Summary:
Internal tp2 glog used with the rest of fbcode has some interesting customizations:
* definition of CRITICAL and VERBOSE logging in addition to the glog default levels
* Some handling of getting / setting thread-local context

We can't really mutate the centos-provided glog (because it's all extracted from the rpm)  - and we should keep it that way!

But instead... we can play games with how we expose the library on the buck graph, and if we're really careful with header paths, namespaces, and symbols, we can make a wrapper library look enough like glog that it works for our needs!

This diff adds a bit of things to rpm_library to extract headers in a specific way so we can optionally rewrite them and expose them how we want. It also adds a little wrapper library that defines some missing or un-exported bits from glog that are used in other fbcode deps.

The net result is that this library applies the few couple necessary customizations on centos-standard glog to get fbcode projects compiling.

Test Plan:
Cathode builds for TEE now, yay
https://www.internalfb.com/buck2/435ad78b-b5f8-45ca-9a24-b97d8a308bb0

Rollback Plan:

Reviewed By: zubivan

Differential Revision: D80632338

fbshipit-source-id: fc06bac223827bb8edfefac9ec937b8539431d80