antlir: Be a bit careful about how we copy files
Summary:
While trying to replicate the tp2 dep structure, I ran into the issue that the `db` project's headers have mode 0444. This combined with the prior way of copying files meant that if we:
1. Copied a directory recursively
2. Tried to copy another file over top of an existing header with mode 0444

Then we'd get a perm denied error because the headers aren't writable

We can't make these writable because the subvol is readonly. Instead, lets partition this into dirs and not dirs. If we copy dirs recursively, we assume that we'll copy over much of the files. Anything we don't, we'll copy over in the second loop.

Test Plan:
db builds now:
https://www.internalfb.com/buck2/0af4a8ab-4045-4219-8c71-65178df64002

Rollback Plan:

Reviewed By: vmagro

Differential Revision: D80682199

fbshipit-source-id: 8420efd6567fd4949478387b37a4cb03b330da68