tee: Mark a few cuda deps as `provided`
Summary:
`provided` is a special attribute of prebuilt_cxx_library's that tells buck this library will be provided by the runtime environment. The net effect is that these .so's won't be bundled into a PAR or included in link groups.

Some cuda libraries - like libcuda and libnvidia-ml - are special stub libraries that link to the DSOs provided by the nvidia driver. At runtime we want to refer to the _real_ DSOs from the driver, but build time still uses the stubs.

Test Plan:
CI

Rollback Plan:

Reviewed By: Rittzz

Differential Revision: D81015191

fbshipit-source-id: 2c4b93f3e1fc9a30b65fd3cd4e429c65883f9ad9